refactor: Mobile Nav toggle to MobileNavMenu  component
